Move random_string method to text module
