+++ temba/orgs/models.py
@@ -37,11 +37,12 @@ from requests import Session
 from smartmin.models import SmartModel
 from temba.bundles import get_brand_bundles, get_bundle_map
 from temba.locations.models import AdminBoundary, BoundaryAlias
-from temba.utils import analytics, str_to_datetime, get_datetime_format, datetime_to_str, random_string, languages
+from temba.utils import analytics, str_to_datetime, get_datetime_format, datetime_to_str, languages
 from temba.utils.cache import get_cacheable_result, get_cacheable_attr, incrby_existing
 from temba.utils.currencies import currency_for_country
 from temba.utils.email import send_template_email, send_simple_email, send_custom_smtp_email
 from temba.utils.models import SquashableModel
+from temba.utils.text import random_string
 from temba.utils.timezones import timezone_to_country_code
 from timezone_field import TimeZoneField
 from urlparse import urlparse
